# docs/swagger.yaml
definitions:
...
  handlers.Reservation:
    properties:
      booking_date:
        type: string
      company:
        type: string
      created_at:
        type: string
      end_time:
        type: string
      final_price:
        type: integer
      id:
        type: string
      name:
        type: string
      participants:
        type: integer
      phone:
        type: string
      room:
        $ref: '#/definitions/handlers.Rooms'
      room_id:
        type: string
      room_price:
        type: integer
      snack:
        $ref: '#/definitions/handlers.Snacks'
      snack_id:
        type: string
      start_time:
        type: string
      status:
        type: string
      total_snack:
        type: integer
      total_snack_price:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/handlers.UserShow'
      user_id:
        type: string
    type: object
  handlers.ReservationInput:
    properties:
      booking_date:
        type: string
      company:
        type: string
      end_time:
        type: string
      name:
        type: string
      participants:
        type: integer
      phone:
        type: string
      room_id:
        example: 6066f8a1-0a80-4299-86ca-99888912bbe5
        type: string
      snack_id:
        example: b8f8cab4-9f0e-4d08-88aa-9fd465a52536
        type: string
      start_time:
        type: string
      total_snack:
        type: integer
      user_id:
        example: 21691490-6817-4bf4-9bf7-3bf624d210a7
        type: string
    type: object
...
  handlers.Rooms:
    properties:
      capacity:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      type:
        type: string
    type: object
  handlers.Snacks:
    properties:
      category:
        type: string
      id:
        type: string
      name:
        type: string
      package:
        type: string
      price:
        type: string
    type: object
  handlers.UpdateStatusInput:
    properties:
      status:
        example: cancel
        type: string
    required:
    - status
    type: object
...
  handlers.UserShow:
    properties:
      email:
        type: string
      id:
        type: string
      image_id:
        type: string
      language:
        type: string
      role:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: This is documentation API from Swagger. Jika ada masalah token silahkan
    coba pakai postman
  title: Swagger MeetApp By Sinau Koding API
  version: "1.0"
paths:
...
  /reservations:
    get:
      consumes:
      - application/json
      description: 'Retrieve all reservations in the system contoh: /reservations?status=cancel&room_type=medium'
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by room type
        in: query
        name: room_type
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Reservation'
            type: array
      security:
      - BearerAuth: []
      summary: Get all reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: 'Note: Untuk Booking date menggunakan format YYYY-MM-DD untuk time
        menggunakan format HH:MM:SS'
      parameters:
      - description: Reservation details
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/handlers.ReservationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new reservation
      tags:
      - reservations
  /reservations/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Reservation'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a reservation by ID
      tags:
      - reservations
    put:
      consumes:
      - application/json
      description: Update the status of a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Reservation status update
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or missing status
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Reservation not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update reservation status
      tags:
      - reservations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"